# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Agent Gateway API
  description: Agent 网关中台后端，提供账户、项目、服务管理和模型网关接口。
  version: 0.1.0
paths:
...
  /api/admin/v1/services:
    get:
      tags:
      - Project Services
      summary: 列出服务目录和模型
      description: 提供当前可申请的服务目录及额度档位。
      operationId: list_service_catalog_api_admin_v1_services_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ServiceCatalogResponse'
                type: array
                title: Response List Service Catalog Api Admin V1 Services Get
      security:
      - HTTPBearer: []
  /api/admin/v1/users/lookup:
    get:
      tags:
      - Project Services
      summary: 按用户名或用户 ID 查找额度目标
      description: 管理员按唯一用户名或 UUID 精确查找用户。
      operationId: lookup_quota_target_api_admin_v1_users_lookup_get
      security:
      - HTTPBearer: []
      parameters:
      - name: user_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          description: 目标用户 UUID；与 username 二选一。
          title: User Id
        description: 目标用户 UUID；与 username 二选一。
      - name: username
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            minLength: 1
            maxLength: 100
          - type: 'null'
          description: 目标用户的精确用户名。
          title: Username
        description: 目标用户的精确用户名。
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuotaTargetResponse'
        '403':
          description: 需要管理员权限
        '404':
          description: 用户不存在
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/admin/v1/users/{user_id}/services:
    get:
      tags:
      - Project Services
      summary: 查看指定用户服务月额度
      description: 管理员查看指定用户的额度、项目分配及跨项目当月用量。
      operationId: get_user_service_quotas_api_admin_v1_users__user_id__services_get
      security:
      - HTTPBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserServiceQuotaResponse'
                title: Response Get User Service Quotas Api Admin V1 Users  User Id  Services
                  Get
        '403':
          description: 需要管理员权限
        '404':
          description: 用户不存在
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/admin/v1/users/{user_id}/services/{service_code}:
    put:
      tags:
      - Project Services
      summary: 设置用户服务月额度
      description: 管理员配置指定用户的服务级月额度。
      operationId: set_user_service_quota_api_admin_v1_users__user_id__services__service_code__put
      security:
      - HTTPBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      - name: service_code
        in: path
        required: true
        schema:
          type: string
          title: Service Code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetUserServiceQuotaRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserServiceQuotaResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/admin/v1/services/{service_code}/projects:
    get:
      tags:
      - Project Services
      summary: 列出已开通指定服务的可访问项目
      description: 返回当前用户可查看且已开通指定服务的运行中项目，管理员可查看全部。
      operationId: list_projects_for_service_api_admin_v1_services__service_code__projects_get
      security:
      - HTTPBearer: []
      parameters:
      - name: service_code
        in: path
        required: true
        schema:
          type: string
          title: Service Code
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceProjectPageResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/admin/v1/projects/{project_id}/services:
    get:
      tags:
      - Project Services
      summary: 查询项目已开通服务和用量
      description: 项目成员、公开项目 review 用户和管理员可以查看；管理员只读。
      operationId: list_project_services_api_admin_v1_projects__project_id__services_get
      security:
      - HTTPBearer: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectServiceResponse'
                title: Response List Project Services Api Admin V1 Projects  Project
                  Id  Services Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Project Services
      summary: 申请项目服务
      description: 项目 owner 提交申请后自动开通，无支付流程。
      operationId: apply_project_service_api_admin_v1_projects__project_id__services_post
      security:
      - HTTPBearer: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceApplicationRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectServiceResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/admin/v1/projects/{project_id}/services/{service_code}:
    patch:
      tags:
      - Project Services
      summary: 调整项目服务额度
      description: 项目 owner 可调整项目额度；执行时校验个人可分配余额和当月已用量。
      operationId: update_project_service_allocation_api_admin_v1_projects__project_id__services__service_code__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      - name: service_code
        in: path
        required: true
        schema:
          type: string
          title: Service Code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectServiceAllocationRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectServiceResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/me/services:
    get:
      tags:
      - My Services
      summary: 查看我的服务和额度
      description: 返回服务目录中当前用户获授的额度、项目分配和当月用量。
      operationId: list_my_services_api_v1_me_services_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/UserServiceQuotaResponse'
                type: array
                title: Response List My Services Api V1 Me Services Get
      security:
      - HTTPBearer: []
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ProjectServiceAllocationRequest:
      properties:
        monthly_token_limit:
          type: integer
          minimum: 1.0
          title: Monthly Token Limit
          description: 新的项目月度 token 上限；不能低于当月已用量或超出 owner 个人上限。
      type: object
      required:
      - monthly_token_limit
      title: ProjectServiceAllocationRequest
    ProjectServiceResponse:
      properties:
        project_id:
          type: string
          format: uuid
          title: Project Id
          description: 订阅所属项目；调用方无需把该 ID 传给模型网关。
        service_code:
          type: string
          title: Service Code
          description: 服务目录代码。
        monthly_token_limit:
          anyOf:
          - type: integer
          - type: 'null'
          title: Monthly Token Limit
          description: AI 服务的每月 token 上限；数据服务为 null。
        status:
          type: string
          enum:
          - active
          - suspended
          title: Status
          description: 服务订阅状态。
        period_start:
          type: string
          title: Period Start
          description: 当前用量周期起始日，UTC 月初。
        tokens_used:
          type: integer
          title: Tokens Used
          description: 本周期已确认消耗。
        tokens_reserved:
          type: integer
          title: Tokens Reserved
          description: 并发请求预留中 token。
      type: object
      required:
      - project_id
      - service_code
      - monthly_token_limit
      - status
      - period_start
      - tokens_used
      - tokens_reserved
      title: ProjectServiceResponse
      description: 项目的有效服务订阅及当前 UTC 月用量。
...
    QuotaTargetResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
          description: 目标用户 ID。
        username:
          type: string
          title: Username
          description: 唯一用户名。
        role:
          type: string
          enum:
          - admin
          - user
          title: Role
          description: 当前平台角色。
      type: object
      required:
      - id
      - username
      - role
      title: QuotaTargetResponse
      description: 最少字段的用户识别结果，禁止返回账户凭据。
...
    ServiceApplicationRequest:
      properties:
        service_code:
          type: string
          title: Service Code
          description: 要申请的服务目录代码。
        monthly_token_limit:
          anyOf:
          - type: integer
            minimum: 1.0
          - type: 'null'
          title: Monthly Token Limit
          description: AI 推理服务的项目月度 token 上限；数据服务不提供此字段。
      type: object
      required:
      - service_code
      title: ServiceApplicationRequest
    ServiceCatalogResponse:
      properties:
        code:
          type: string
          title: Code
          description: 稳定服务代码。
        name:
          type: string
          title: Name
          description: 服务展示名称。
        models:
          items:
            type: string
          type: array
          title: Models
          description: 此服务包含的模型标识。
        quota_unit:
          anyOf:
          - type: string
            const: tokens
          - type: 'null'
          title: Quota Unit
          description: AI 推理服务的计量单位；null 表示数据服务不占用 AI 推理额度。
      type: object
      required:
      - code
      - name
      - models
      - quota_unit
      title: ServiceCatalogResponse
      description: 目前可申请的服务目录项。
    ServiceProjectPageResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ServiceProjectResponse'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        offset:
          type: integer
          title: Offset
        limit:
          type: integer
          title: Limit
      type: object
      required:
      - items
      - total
      - offset
      - limit
      title: ServiceProjectPageResponse
    ServiceProjectResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        visibility:
          type: string
          enum:
          - public
          - private
          title: Visibility
        owner_id:
          type: string
          format: uuid
          title: Owner Id
        service_code:
          type: string
          title: Service Code
        monthly_token_limit:
          anyOf:
          - type: integer
          - type: 'null'
          title: Monthly Token Limit
        service_status:
          type: string
          enum:
          - active
          - suspended
          title: Service Status
        tokens_used:
          type: integer
          title: Tokens Used
        tokens_reserved:
          type: integer
          title: Tokens Reserved
      type: object
      required:
      - id
      - name
      - description
      - visibility
      - owner_id
      - service_code
      - monthly_token_limit
      - service_status
      - tokens_used
      - tokens_reserved
      title: ServiceProjectResponse
      description: 当前用户可访问且已开通指定服务的项目。
...
    SetUserServiceQuotaRequest:
      properties:
        monthly_token_limit:
          type: integer
          minimum: 0.0
          title: Monthly Token Limit
          description: 用户每月服务总上限；允许低于项目分配总额，0 会暂停调用且保留项目分配记录。
      type: object
      required:
      - monthly_token_limit
      title: SetUserServiceQuotaRequest
...
    UserServiceQuotaResponse:
      properties:
        service_code:
          type: string
          title: Service Code
          description: 服务目录代码。
        name:
          type: string
          title: Name
          description: 服务展示名称。
        models:
          items:
            type: string
          type: array
          title: Models
          description: 服务包含的模型。
        monthly_token_limit:
          type: integer
          title: Monthly Token Limit
          description: 管理员授予的用户月度总上限；0 表示服务调用暂停，项目分配记录仍可保留。
        allocated_tokens:
          type: integer
          title: Allocated Tokens
          description: 已分配到该用户所有项目的额度合计。
        available_tokens:
          type: integer
          title: Available Tokens
          description: 还能分配给项目的额度。
        tokens_used:
          type: integer
          title: Tokens Used
          description: 该用户所有项目本月实际消耗合计。
        tokens_reserved:
          type: integer
          title: Tokens Reserved
          description: 该用户所有项目当前请求预留合计。
      type: object
      required:
      - service_code
      - name
      - models
      - monthly_token_limit
      - allocated_tokens
      - available_tokens
      - tokens_used
      - tokens_reserved
      title: UserServiceQuotaResponse
      description: 个人服务能力和跨项目额度分配概览。
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer